Fixed a cast in string_utils.h

+++ source/tools/string_utils.h
@@ -41,7 +41,7 @@ namespace emp {
   /// Count the number of times a specific character appears in a string
   /// (a clean shortcut to std::count)
   static inline size_t count(const std::string & str, char c) {
-    return std::count(str.begin(), str.end(), c);
+    return (size_t) std::count(str.begin(), str.end(), c);
   }
 
   /// Convert a single chararcter to one that uses a proper escape sequence (in a string) if needed.
